python-test-samples/apigw-lambda-dynamodb/tests/requirements.txt
pytest
pytest-mock
boto3
moto
aws_lambda_powertools
aws-xray-sdk
fastjsonschema
pyyaml